UPdate -> AudioManager + CoinSystem
cy
