--- Assets/CSE5912/Sound/AudioManager.cs
@@ -1,30 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class AudioManager : MonoBehaviour
-{
-    //Sound is a public class build for manager soundeffect.
-    public Sound[] sounds;
-
-    private void Awake()
-    {
-        foreach(Sound s in sounds)
-        {
-            s.source = gameObject.AddComponent<AudioSource>();
-            s.source.clip = s.clip;
-
-            s.source.volume = s.volume;
-            s.source.pitch = s.pitch;
-        }
-    }
-
-    public void Play (string name)
-    {
-        Sound s = Array.Find(sounds, sound => sound.name == name);
-        s.source.Play(); 
-    }
-    //How to use it.
-    //FindObjectOfType<AudioManager>().Play("GetAttack");
-}
