--- Assets/CSE5912/Sound/AudioManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 120d54dc54596f245be16aba006e8129
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
