--- Assets/CSE5912/Sound/PlaySoundOnStart.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a06baa1e26bce14dbce0896fd6710e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
