+++ Assets/CSE5912/Sound/Sound.cs
@@ -11,8 +11,11 @@ public class Sound {
     [Range(0f, 1f)]
     public float volume;
 
+    //a quality that makes a melody go higher or lower.
     [Range(.1f, 3f)]
     public float pitch;
+
+    public bool loop;
 
     [HideInInspector]
     public AudioSource source;
