+++ Assets/CSE5912/Sound/SoundManager.cs
@@ -1,28 +1,56 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public class SoundManager : MonoBehaviour
 {
-    public static SoundManager Instance;
+    //Sound is a public class build for manager soundeffect.
+    public Sound[] sounds;
 
-    [SerializeField] private AudioSource  _effectSource,_musicSource;
+    public static SoundManager instance;
 
     private void Awake()
     {
-        if(Instance = null)
+/*        if (instance = null)
         {
-            Instance = this;
-            DontDestroyOnLoad(gameObject);
+            instance = this;
         }
         else
         {
             Destroy(gameObject);
+            return;
+        }
+        DontDestroyOnLoad(gameObject);*/
+
+        foreach (Sound s in sounds)
+        {
+            s.source = gameObject.AddComponent<AudioSource>();
+            s.source.clip = s.clip;
+
+            s.source.volume = s.volume;
+            s.source.pitch = s.pitch;
+            s.source.loop = s.loop;
         }
     }
 
-    public void PlayEffect(AudioClip clip)
+    //Use for later. Background music
+    private void Start()
     {
-        _effectSource.PlayOneShot(clip);
+        Play("BackgroundMusic");
     }
+
+    public void Play (string name)
+    {
+        Sound s = Array.Find(sounds, sound => sound.name == name);
+        if(s == null)
+        {
+            Debug.LogWarning("Sound " + name + " is not founded.");
+            return;
+        }
+        s.source.Play();
+        Debug.Log("Play sound " + name);
+    }
+    //How to use it.
+    //FindObjectOfType<SoundManager>().Play("");
 }
